requirements_dev.txt
-r requirements.txt

django-debug-toolbar>=1.3.2,<1.4
